# docs-website/static/docs/api/v1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Documentation for the traduora REST API\n\nOfficial website: https://traduora.com\nAdditional documentation: https://docs.traduora.com\nSource code: https://github.com/traduora/traduora",
    "version": "0.14.0",
    "title": "Traduora API"
  },
  "basePath": "/",
...
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "oauth2": {
      "type": "oauth2",
      "flow": "password",
      "authorizationUrl": "/api/v1/auth/token",
      "tokenUrl": "/api/v1/auth/token"
    }
  },
  "paths": {
...
    "/api/v1/projects/{projectId}": {
      "get": {
        "summary": "Get a project by id",
        "parameters": [
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ProjectResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["Projects"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "patch": {
        "summary": "Update a project",
        "parameters": [
          {
            "name": "UpdateProjectRequest",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateProjectRequest"
            }
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ProjectResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["Projects"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "delete": {
        "summary": "Delete a project",
        "parameters": [
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["Projects"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      }
    },
...
    "/api/v1/projects/{projectId}/terms/{termId}": {
      "patch": {
        "summary": "Update a project's term",
        "parameters": [
          {
            "name": "UpdateTermRequest",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateTermRequest"
            }
          },
          {
            "type": "string",
            "name": "termId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/ProjectTermResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Terms"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "delete": {
        "summary": "Remove a project's term",
        "description": "Removes a project's term and all related translations",
        "parameters": [
          {
            "type": "string",
            "name": "termId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Terms"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      }
    },
...
    "/api/v1/projects/{projectId}/labels": {
      "get": {
        "summary": "List a project's labels",
        "parameters": [
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ProjectLabelResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Project Labels"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "post": {
        "summary": "Add a new project label",
        "parameters": [
          {
            "name": "AddLabelRequest",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AddLabelRequest"
            }
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ProjectLabelResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Plan limit reached"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Project Labels"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      }
    },
    "/api/v1/projects/{projectId}/labels/{labelId}": {
      "patch": {
        "summary": "Update a project's label",
        "parameters": [
          {
            "name": "UpdateLabelRequest",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateLabelRequest"
            }
          },
          {
            "type": "string",
            "name": "labelId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/ProjectLabelResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Project Labels"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "delete": {
        "summary": "Remove a project's label",
        "description": "Removes a project's label and removes it from all attached translations too",
        "parameters": [
          {
            "type": "string",
            "name": "labelId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Project Labels"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      }
    },
    "/api/v1/projects/{projectId}/labels/{labelId}/terms/{termId}": {
      "post": {
        "summary": "Label a project term",
        "parameters": [
          {
            "type": "string",
            "name": "termId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "labelId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project, term or label not found"
          }
        },
        "tags": ["Project Labels"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "delete": {
        "summary": "Unlabel a project term",
        "parameters": [
          {
            "type": "string",
            "name": "termId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "labelId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Label removed"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project, term or label not found"
          }
        },
        "tags": ["Project Labels"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      }
    },
    "/api/v1/projects/{projectId}/labels/{labelId}/terms/{termId}/translations/{localeCode}": {
      "post": {
        "summary": "Label a project term translation",
        "parameters": [
          {
            "type": "string",
            "name": "localeCode",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "termId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "labelId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project, term or label not found"
          }
        },
        "tags": ["Project Labels"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "delete": {
        "summary": "Unlabel a project term translation",
        "parameters": [
          {
            "type": "string",
            "name": "localeCode",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "termId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "labelId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Label removed"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project, term or label not found"
          }
        },
        "tags": ["Project Labels"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      }
    },
...
  },
  "definitions": {
...
    "AccessDatesDTO": {
      "type": "object",
      "properties": {
        "created": {
          "type": "string"
        },
        "modified": {
          "type": "string"
        }
      },
      "required": ["created", "modified"]
    },
    "ProjectDTO": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "localesCount": {
          "type": "number"
        },
        "termsCount": {
          "type": "number"
        },
        "role": {
          "type": "string",
          "enum": ["admin", "editor", "viewer"]
        },
        "date": {
          "$ref": "#/definitions/AccessDatesDTO"
        }
      },
      "required": ["id", "name", "description", "localesCount", "termsCount", "role", "date"]
    },
...
    "ProjectResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/ProjectDTO"
        }
      },
      "required": ["data"]
    },
    "UpdateProjectRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        },
        "description": {
          "type": "string",
          "minLength": 0,
          "maxLength": 255
        }
      }
    },
...
    "ProjectTermDTO": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "labels": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "date": {
          "$ref": "#/definitions/AccessDatesDTO"
        }
      },
      "required": ["id", "value", "labels", "date"]
    },
...
    "ProjectTermResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/ProjectTermDTO"
        }
      },
      "required": ["data"]
    },
    "UpdateTermRequest": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        }
      },
      "required": ["value"]
    },
...
    "ProjectLabelDTO": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "color": {
          "type": "string"
        }
      },
      "required": ["id", "value", "color"]
    },
    "ProjectLabelResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/ProjectLabelDTO"
        }
      },
      "required": ["data"]
    },
    "AddLabelRequest": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        },
        "color": {
          "type": "string",
          "minLength": 7,
          "maxLength": 7
        }
      },
      "required": ["value", "color"]
    },
    "UpdateLabelRequest": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        },
        "color": {
          "type": "string",
          "minLength": 7,
          "maxLength": 7
        }
      },
      "required": ["value", "color"]
    },
...
  }
}